@@ -1,10 +1,10 @@
 # SPDX-License-Identifier: MIT
 
+from __future__ import annotations
+
 import argparse
 import sys
 import timeit
-
-from typing import List
 
 from . import (
     DEFAULT_HASH_LENGTH,
@@ -16,7 +16,7 @@ from . import (
 )
 
 
-def main(argv: List[str]) -> None:
+def main(argv: list[str]) -> None:
     parser = argparse.ArgumentParser(description="Benchmark Argon2.")
     parser.add_argument(
         "-n", type=int, default=100, help="Number of iterations to measure."
